fix: improve drag swap precision and add tests
- Update `DraggableLazyColumn` to use float arithmetic for accurate item swap detection
- Add `DraggableLazyColumnTest` covering drag-down, drag-up, and cancellation scenarios
- Update IDE configuration files for Android test results and deployment targets

+++ app/src/main/java/com/mrkongtk/rtspviewer/ui/compose/DraggableLazyColumn.kt
@@ -136,7 +136,7 @@ fun <T> DraggableLazyColumn(
                                     .lastOrNull { it.index == draggingItemIndex - 1 }
 
                                 // Check if we should swap DOWN
-                                if (itemBelow != null && currentBottom > itemBelow.offset + (itemBelow.size / 2)) {
+                                if (itemBelow != null && currentBottom > itemBelow.offset.toFloat() + (itemBelow.size.toFloat() / 2f)) {
                                     // Move data in the list
                                     items.add(
                                         draggingItemIndex + 1,
@@ -151,7 +151,7 @@ fun <T> DraggableLazyColumn(
                                     draggingItemOffset -= itemBelow.size
                                 }
                                 // Check if we should swap UP
-                                else if (itemAbove != null && currentTop < itemAbove.offset + (itemAbove.size / 2)) {
+                                else if (itemAbove != null && currentTop < itemAbove.offset.toFloat() + (itemAbove.size.toFloat() / 2f)) {
                                     // Move data in the list
                                     items.add(
                                         draggingItemIndex - 1,
